Refacto familyVariationAxeTask and fix scope in akeneoAttributeDataProvider
